% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{annotate}
\alias{annotate}
\title{Run the annotation pipeline on a set of documents}
\usage{
annotate(input, output_dir = NULL, load = TRUE, keep = TRUE,
  as_strings = FALSE, doc_id_offset = 0L)
}
\arguments{
\item{input}{either a vector of file names to parse, or a character vector
with one document in each element. Specify the latter with the
as_string flag.}

\item{output_dir}{path to the directory where the output should be stored. Will be
created if it does not exist. Files currently in this location will
be overwritten. If NULL, the default, it uses a temporary directory.}

\item{load}{logical. Once parsed, should the data be read into R as an annotation object}

\item{keep}{logical. Once parsed, should the files be kept on disk.}

\item{as_strings}{logical. Is the data given to \code{input} the actual document text rather
file names.}

\item{doc_id_offset}{integer. The first document id to use. Defaults to 0.}
}
\description{
Run the annotation pipeline on a set of documents
}

