% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{cnlp_annotate}
\alias{cnlp_annotate}
\title{Run the annotation pipeline on a set of documents}
\usage{
cnlp_annotate(input, backend = NULL, verbose = TRUE)
}
\arguments{
\item{input}{a data frame containing the data to parse. Must
contain a column called 'text' (with the raw text to
parse) or a column called 'path' (a path to the file
that should be parsed). If both are given, the
function defers to the text column.}

\item{backend}{name of the backend to use. Will default to the last
model to be initalized.}

\item{verbose}{logical; should annotation engine print out when it
finishes each text? Turned on by default.}
}
\value{
an object of class \code{annotation}
}
\description{
Runs the clean_nlp annotators over a given corpus of text
using the desired backend. The details for which annotators to run and
how to run them are specified by using one of:
\code{\link{cnlp_init_stringi}}, \code{\link{cnlp_init_spacy}},
\code{\link{cnlp_init_udpipe}}, or \code{\link{cnlp_init_corenlp}}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
