% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif.R
\name{cnlp_annotate_tif}
\alias{cnlp_annotate_tif}
\title{Wrapper for TIF Compliant Usage}
\usage{
cnlp_annotate_tif(input)
}
\arguments{
\item{input}{a data frame representing a tif corpus object.
The first column should contain a unique document
identifier and the second should contain the raw
text.}
}
\value{
an annotation object
}
\description{
Runs annotations over a corpus object and returns results as
an annotation object. See parameters below for the specific
input format For more control, see the function
\code{\link{cnlp_annotate}}.
}
\examples{
\dontrun{
annotation <- cnlp_annotate_tif(tif_data_frame)
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
