% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_tools.R
\name{cnlp_combine_documents}
\alias{cnlp_combine_documents}
\title{Combine a set of annotations}
\usage{
cnlp_combine_documents(...)
}
\arguments{
\item{...}{annotation objects to combine; either a single list item
or all of the objects as individual inputs}
}
\value{
a single annotation object containing all of the input documents
}
\description{
Takes an arbitrary set of annotations and efficiently combines them into
a single object. All document ids are reset so that they are contiguous
integers starting at zero.
}
\examples{
\dontrun{
annotation <- combine_annotators(anno01, anno02, anno03)
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
