% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{cnlp_download_corenlp}
\alias{cnlp_download_corenlp}
\title{Download model files needed for coreNLP}
\usage{
cnlp_download_corenlp(lang = "en")
}
\arguments{
\item{lang}{string giving the languange code. Defaults to "en".}
}
\description{
The cleanNLP package does not supply the model files required
for using the coreNLP backend. These files can be downloaded
with this function. If you need more control, download directly from
Python.
}
\examples{
\dontrun{
cnlp_download_corenlp(lang="en")
cnlp_download_corenlp(lang="de")
}
}
