% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{cnlp_download_spacy}
\alias{cnlp_download_spacy}
\title{Download model files needed for spacy}
\usage{
cnlp_download_spacy(model_name = "en")
}
\arguments{
\item{model_name}{string giving the model namel.
Defaults to "en".}
}
\description{
The cleanNLP package does not supply the model files required
for using the spacy backend. These files can be downloaded
with this function. If you need more control, download directly from
Python.
}
\examples{
\dontrun{
cnlp_download_spacy(model_name="en")
cnlp_download_spacy(model_name="de")
}
}
