% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{cnlp_download_udpipe}
\alias{cnlp_download_udpipe}
\title{Download model files needed for udpipe}
\usage{
cnlp_download_udpipe(model_name = "english", model_loc = NULL)
}
\arguments{
\item{model_name}{string giving the model namel.
Defaults to "english" if NULL.}

\item{model_loc}{where should be model be downloaded to. If
set to NULL, will be downloaded in the location
that the cleanNLP package is installed.}
}
\description{
The cleanNLP package does not supply the model files required
for using the udpipe backend. These files can be downloaded
with this function. The models are saved, by default, in the
location where the package is installed. They will be saved
between R sessions. This function is called internally by
\code{\link{cnlp_init_udpipe}} if a model is not available.
}
\examples{
\dontrun{
cnlp_download_core_nlp()
cnlp_download_core_nlp(type="spanish")
}
}
