% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{cnlp_get_vector}
\alias{cnlp_get_vector}
\title{Access word embedding vector from an annotation object}
\usage{
cnlp_get_vector(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\value{
Returns a matrix containing one row for every triple found
 in the corpus, or \code{NULL} if not embeddings are present
}
\description{
Word embeddings map each lemma or token into a high-dimensional
vector space. The implementation here uses a 300-dimensional
space. Only available with the spaCy parser.
}
\references{
Pennington, Jeffrey, Richard Socher, and Christopher D. Manning.
"Glove: Global Vectors for Word Representation." EMNLP. Vol. 14. 2014.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
