% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_corenlp.R
\name{cnlp_init_corenlp_custom}
\alias{cnlp_init_corenlp_custom}
\title{Interface for initializing the corenlp backend with custom choices of 
annotators and settings}
\usage{
cnlp_init_corenlp_custom(language, lib_location = NULL, mem = "6g",
  verbose = FALSE, keys, values)
}
\arguments{
\item{language}{a character vector describing the desired language;
should be one of: "ar", "de", "en", "es", "fr",
or "zh".}

\item{lib_location}{a string giving the location of the corenlp java
files. This should point to a directory which
contains, for example the file
"stanford-corenlp-*.jar",
where "*" is the version number. If missing, the
function will try to find the library in the
environment variable corenlp_HOME, and otherwise
will fail. (Java model only)}

\item{mem}{a string giving the amount of memory to be assigned
to the rJava engine. For example, "6g" assigned 6
gigabytes of memory. At least 2 gigabytes are
recommended at a minimum for running the corenlp
package. On a 32bit machine, where this is not
possible, setting "1800m" may also work. This
option will only have an effect the first
time \code{init_backend} is called for the corenlp
backend, and also will not have an effect if the
java engine is already started by another process.}

\item{verbose}{boolean. Should messages from the pipeline be
written to the console or suppressed?}

\item{keys}{vector string of flags to add to the corenlp calls}

\item{values}{vector string of values paired with the flags.

The example below shows how to initialise corenlp to run named entity 
recognition (ner) with its respective dependencies (tokenize, ssplit, pos 
and lemma). More details of flag (key) options and their values at 
\url{https://stanfordnlp.github.io/CoreNLP/}}
}
\description{
As an alternative to cnlp_init_corenlp, this function must be run before
annotating text with the corenlp backend with custom options. It sets the
properties for the corenlp engine and loads the file using rJava interface
provided by reticulate.
}
\examples{
\dontrun{
cnlp_init_corenlp_custom(language = "en", mem = "2g",
                         keys="annotators", values="tokenize, ssplit, pos, lemma, ner")
}

}
