% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cnlp_init_stringi}
\alias{cnlp_init_stringi}
\title{Interface for initializing the standard R backend}
\usage{
cnlp_init_stringi(locale = NULL)
}
\arguments{
\item{locale}{string giving the locale name to
pass to the stringi functions. If
\code{NULL}, the default locale is
selected}
}
\description{
This function must be run before annotating text with
the tokenizers backend.
}
\examples{
\dontrun{
cnlp_init_tokenizers()
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
