% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{cnlp_quick}
\alias{cnlp_quick}
\title{Quickly Compute Data Frame of Annotations}
\usage{
cnlp_quick(input, ...)
}
\arguments{
\item{input}{either a vector of file names to parse, a
character vector with one document in each
element, or a data frame. If a data frame,
specify what column names contain the text and
(optionally) document ids}

\item{...}{additional options passed to
\code{\link{cnlp_annotate}}}
}
\value{
a data frame of annotations with one row per token
}
\description{
Runs the clean_nlp annotators over a given corpus of text
and returns a data frame of annotated text with one token
per line. By default it will initalize the udpipe backend
if no annotators are found.
}
\examples{
\dontrun{
annotation <- cnlp_quick(c("Parse this text.", "This too, as a new doc."))
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
