% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{cnlp_read_csv}
\alias{cnlp_read_csv}
\title{Read annotation files from disk}
\usage{
cnlp_read_csv(input_dir)
}
\arguments{
\item{input_dir}{path to the directory where the files
are stored}
}
\value{
an object of class \code{annotation}
}
\description{
Loads an annotation that has been stored as a set of csv files in a
local directory. This is typically created by a call to
\code{\link{cnlp_annotate}} or \code{\link{cnlp_write_csv}}.
}
\examples{
\dontrun{
annotation <- cnlp_read_csv("path/to/annotation")
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
