% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cnlp_utils_pca}
\alias{cnlp_utils_pca}
\title{Compute Principal Components and store as a Data Frame}
\usage{
cnlp_utils_pca(x, k = 2, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a matrix object to pass to \code{prcomp}}

\item{k}{integer. The number of components to include in the output.}

\item{center}{logical. Should the data be centered?}

\item{scale}{logical. Should the data be scaled? Note that this will
need to be set to false if any columns in \code{x} are
constant if \code{center} is also true.}
}
\value{
a data frame object containing the top \code{k} principal
        components of the data in x.
}
\description{
Takes a matrix and returns a data frame with the top principal components
extracted. This is a simple but powerful technique for visualizing a corpus
of documents.
}
