% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{combine_annotators}
\alias{combine_annotators}
\title{Combine a set of annotations}
\usage{
combine_annotators(...)
}
\arguments{
\item{...}{annotation objects to combine}
}
\description{
Combine a set of annotations
}

