% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dep_frequency}
\alias{dep_frequency}
\title{Universal Dependency Frequencies}
\description{
The language-specific frequency of universal dependency codes as given
in standard treebank corpora. Frequencies are multiplied by 100.
}
\references{
\url{http://universaldependencies.org/}
}
\keyword{data}

