% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{doc_id_reset}
\alias{doc_id_reset}
\title{Reset document ids}
\usage{
doc_id_reset(annotation, start_id = 0L)
}
\arguments{
\item{annotation}{annotation object to reset the IDs of}

\item{start_id}{the starting document id. Defaults to 0.}
}
\description{
Reset document ids
}

