% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_clean_nlp}
\alias{download_clean_nlp}
\title{Download java files needed for cleanNLP}
\usage{
download_clean_nlp(output_loc, type = c("base", "chinese", "english",
  "french", "german", "spanish"))
}
\arguments{
\item{output_loc}{a string showing where the files are to be downloaded.
If missing, will try to download files into the directory
where the package was original installed.}

\item{type}{type of files to download. The base backage, installed by
default, is required. Other jars include chinese, german,
and spanish. These can be installed in addition to the
base package.}
}
\description{
The cleanNLP package does not supply the raw java files
provided by the Stanford NLP Group as they are quite large.
This function downloads the libraries automatically, by default into
the directory where the package was installed.
}
\details{
In order to manually download files, simply unzip them and
specify their location in your call to \code{\link{init_clean_nlp}}
using the \code{lib_location} parameter.
}
\examples{
\dontrun{
download_clean_nlp()
download_clean_nlp(type="spanish")
}
}

