% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadCleanNLP}
\alias{downloadCleanNLP}
\title{Download java files needed for cleanNLP}
\usage{
downloadCleanNLP(outputLoc, type = c("base", "chinese", "english", "french",
  "german", "spanish"))
}
\arguments{
\item{outputLoc}{a string showing where the files are to be downloaded.
If missing, will try to download files into the directory
where the package was original installed.}

\item{type}{type of files to download. The base backage, installed by
default is required. Other jars include chinese, german,
and spanish. These will be installed in addition to the
base package.}
}
\description{
The cleanNLP package does not supply the raw java files
provided by the Stanford NLP Group as they are quite large.
This function downloads the libraries for you, by default into
the directory where the package was installed.
}
\details{
If you want to manually download files, simply unzip them and
place in \code{system.file("extdata", package="cleanNLP")}
}
\examples{
\dontrun{
downloadCleanNLP()
downloadCleanNLP(type="spanish")
}
}

