% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{extract_documents}
\alias{extract_documents}
\title{Extract documents from an annotation object}
\usage{
extract_documents(annotation, ids)
}
\arguments{
\item{annotation}{the object to extract from}

\item{ids}{a vector of integer ids from which to extract}
}
\value{
a new annotation object
}
\description{
Takes an annotation object and returns an annotation object containing only
a subset of the documents.
}
\examples{
\dontrun{
annotation <- extract_documents(anno, ids = c(0, 1, 2, 3))
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
