% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_tools.R
\name{from_CoNLL}
\alias{from_CoNLL}
\title{Reads a CoNLL-U or CoNLL-X File}
\usage{
from_CoNLL(file)
}
\arguments{
\item{file}{character vector giving the path to the file}
}
\value{
an annotation object with a single document
}
\description{
Takes an file saved in the CoNLL-U or CoNLL-X format and converts it
into an annotation object. This is a lossy procedure, grabbing just
tokenization, lemmatization, part of speech tags, and dependencies.
}
\examples{
\dontrun{
from_CoNLL(annotation, "/path/to/file.conll")
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
