% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_annotation}
\alias{get_annotation}
\title{Access annotation metadata from an annotation object}
\usage{
get_annotation(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Access annotation metadata from an annotation object
}

