% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_combine}
\alias{get_combine}
\title{One Table Summary of an Annotation Object}
\usage{
get_combine(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
This function pulls together several tables to provide
a one table, denormalized version of the annotation object.
}
\references{
Manning, Christopher D., Mihai Surdeanu, John Bauer, Jenny Finkel,
  Steven J. Bethard, and David McClosky. 2014.
  \href{http://nlp.stanford.edu/pubs/StanfordCoreNlp2014.pdf}{
  The Stanford CoreNLP Natural Language Processing Toolkit}.
  In: \emph{Proceedings of the 52nd Annual Meeting of the
  Association for Computational Linguistics: System Demonstrations,
  pp. 55-60.}

  Kristina Toutanova and Christopher D. Manning. 2000. Enriching
  the Knowledge Sources Used in a Maximum Entropy Part-of-Speech
  Tagger. In: \emph{Proceedings of the Joint SIGDAT Conference on
  Empirical Methods in Natural Language Processing and Very Large
  Corpora (EMNLP/VLC-2000), pp. 63-70}.

  Kristina Toutanova, Dan Klein, Christopher Manning, and Yoram
  Singer. 2003. Feature-Rich Part-of-Speech Tagging with a Cyclic
  Dependency Network. In: \emph{Proceedings of HLT-NAACL 2003,
  pp. 252-259.}
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
