% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_coreference}
\alias{get_coreference}
\title{Access coreferences from an annotation object}
\usage{
get_coreference(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Access coreferences from an annotation object
}

