% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_dependency}
\alias{get_dependency}
\title{Access dependencies from an annotation object}
\usage{
get_dependency(annotation, get_token = FALSE)
}
\arguments{
\item{annotation}{an annotation object}

\item{get_token}{logical. Should words and lemmas be attachend to the
returned dependency table.}
}
\description{
Access dependencies from an annotation object
}

