% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_document}
\alias{get_document}
\title{Access document metadata from an annotation object}
\usage{
get_document(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Access document metadata from an annotation object
}

