% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_entity}
\alias{get_entity}
\title{Access named entities from an annotation object}
\usage{
get_entity(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Access named entities from an annotation object
}

