% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_sentiment}
\alias{get_sentiment}
\title{Access sentiment scores from an annotation object}
\usage{
get_sentiment(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Access sentiment scores from an annotation object
}

