% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_token}
\alias{get_token}
\title{Access tokens from an annotation object}
\usage{
get_token(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Access tokens from an annotation object
}

