% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_triple}
\alias{get_triple}
\title{Access triples from an annotation object}
\usage{
get_triple(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
Access triples from an annotation object
}

