% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{init_backend}
\alias{init_backend}
\title{Initialize a cleanNLP backend}
\usage{
init_backend(type)
}
\arguments{
\item{type}{name of the backend to initialize}
}
\description{
This should be run prior to calling the `annotation`
function. Options that control the behavior of the
pipeline must be set using one of: \code{\link{setup_tokenizers_backend}},
\code{\link{setup_spaCy_backend}}, or \code{\link{setup_coreNLP_backend}}.
Options may be reset during a given R session; in order to
take effect, make a new call to this function.
}
\examples{
\dontrun{
setup_spaCy_backend()
init_backend(type = "spaCy")
}
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
