% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{init_clean_nlp}
\alias{init_clean_nlp}
\title{Initialize the cleanNLP java object}
\usage{
init_clean_nlp()
}
\description{
This must be run prior to calling any other cleanNLP
functions. Options that control the behavior of the
pipeline must be set using \code{\link{set_cleanNLP}}.
You may reset the options and call this function multiple
times in order to specify a different set of parameter.
}

