% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_tokenizers.R
\name{init_tokenizers}
\alias{init_tokenizers}
\title{Interface for initializing the tokenizers backend}
\usage{
init_tokenizers()
}
\description{
This function must be run before annotating text with
the tokenizers backend.
}
\examples{
\dontrun{
init_spaCy_tokenizers()
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
