% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner_annotate.R
\name{NERAnnotate}
\alias{NERAnnotate}
\title{Run the annotation pipeline on a set of documents to extract entities}
\usage{
NERAnnotate(input.file)
}
\arguments{
\item{input.file}{a character string showing the path to the file to be processed}
}
\value{
a data.frame with the details of the detected entities.
}
\description{
Runs the entity detection algorithms from CoreNLP using either the rJava.
It initializes the CoreNLP Java object with the NER annotation parameters
and a path to a temp file that is used for processing using \code{initCoreNLPNER}.
The function returns a \code{data.frame} showing the location in the document 
where the entity occurs andthe entity type. If no entities are detected for a document 
then a row of NA values is returned.
}
