% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obama}
\alias{obama}
\title{Annotation of Barack Obama's State of the Union Addresses}
\description{
Parsed text from all eight State of the Union addresses given by
Barack Obama.
}
\references{
\url{http://www.presidency.ucsb.edu/sou.php}
}
\keyword{data}
