% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner_annotate.R
\name{parseNEROutput}
\alias{parseNEROutput}
\title{Parse the NER output from CoreNLP}
\usage{
parseNEROutput(file, entity.mentions.only)
}
\arguments{
\item{file}{Character string with the path of the input file without extension to be processed}

\item{entity.mentions.only}{Logical to determine if only the entity mention output is processed. 
The entity mention output includes entities and any grouped words e.g. John Smith is a single entity
instead of two separate entities. However, this output also includes personal pronouns (he, she etc)
as entities. If this logical is set to FALSE, the entity output is validated agains the token output.
The token output specifies if a word is an entity and doesn't classify personal pronouns as entities.}
}
\value{
data.frame with three columns, 
\itemize{
       \item \code{id} integer: the row index of the input file that has an extracted entity.
       \item \code{entity} character: The extracted entity word (e.g. William)
       \item \code{entity.tyoe} character: The entity type of the extracted entity (e.g. Person)
   }
}
\description{
Reads the JSON file output from coreNLP ner annotator and coverts to a data frame
}
