% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pos_frequency}
\alias{pos_frequency}
\title{Universal Part of Speech Code Frequencies}
\description{
The language-specific frequency of universal part of speech codes as given
in standard treebank corpora. Frequencies are multiplied by 100.
}
\references{
\url{http://universaldependencies.org/}
}
\keyword{data}

