% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{print.annotation}
\alias{print.annotation}
\title{Print a summary of an annotation object}
\usage{
\method{print}{annotation}(x, ...)
}
\arguments{
\item{x}{an annotation object}

\item{...}{other arguments. Currently unused.}
}
\description{
Print a summary of an annotation object
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
