% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{read_annotation}
\alias{read_annotation}
\title{Read annotation files from disk}
\usage{
read_annotation(input_dir)
}
\arguments{
\item{input_dir}{path to the directory where the files are stored}
}
\description{
Read annotation files from disk
}

