% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed.R
\name{renamed}
\alias{renamed}
\alias{get_combine}
\alias{get_coreference}
\alias{get_dependency}
\alias{get_document}
\alias{get_entity}
\alias{get_sentence}
\alias{get_tfidf}
\alias{get_token}
\alias{get_vector}
\alias{init_coreNLP}
\alias{init_spaCy}
\alias{init_tokenizers}
\alias{run_annotators}
\alias{tidy_pca}
\alias{to_CoNNL}
\alias{from_CoNNL}
\alias{write_annotation}
\alias{read_annotation}
\alias{download_core_nlp}
\alias{combine_documents}
\alias{extract_documents}
\title{Renamed functions}
\usage{
get_combine(...)

get_coreference(...)

get_dependency(...)

get_document(...)

get_entity(...)

get_sentence(...)

get_tfidf(...)

get_token(...)

get_vector(...)

init_coreNLP(...)

init_spaCy(...)

init_tokenizers(...)

run_annotators(...)

tidy_pca(...)

to_CoNNL(...)

from_CoNNL(...)

write_annotation(...)

read_annotation(...)

download_core_nlp(...)

combine_documents(...)

extract_documents(...)
}
\arguments{
\item{...}{options passed to the newly named function}
}
\description{
These functions have been renamed. For the most
part they should now just be called with the prefix
'cnlp_'. See individual warning messages for the
particular calling structure.
}
