% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{save_annotation}
\alias{save_annotation}
\title{Write annotation files to disk}
\usage{
save_annotation(annotation, output_dir)
}
\arguments{
\item{annotation}{annotation file being stored}

\item{output_dir}{path to the directory where the files will be saved}
}
\description{
Write annotation files to disk
}

