% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{set_annotators}
\alias{set_annotators}
\title{Wrapper to set particular annotators}
\usage{
set_annotators(annotators = NULL)
}
\arguments{
\item{annotators}{a character vector describing the desired annotators;
should be a subset of: "segment", "tokenize", "ssplit", "pos", "lemma",
"ner", "parse", "mention", "dcoref", "natlog", "openie",
and "sentiment"}
}
\description{
This function calls \code{\link{set_properties}}, setting the correct set of
annotators that will be used in the pipeline. This function makes sure that
for whichever annotators are choose, all required prerequisites are also included.
Generally, this function will be of most use for parsing English. See the function
\code{\link{set_language}} for a more user-friendly approach.
}
\examples{
\dontrun{
set_annotators(c("segment", "tokenize", "ssplit", "pos", "lemma"))
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}

