% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{set_cleanNLP}
\alias{set_cleanNLP}
\title{Store properties used when initalizing the CoreNLP pipeline}
\usage{
set_cleanNLP(type = c("english", "english_fast", "english_very_fast",
  "arabic", "chinese", "french", "german", "spanish"), libLoc = NULL,
  mem = "4g")
}
\arguments{
\item{type}{type of model to load. See details for options.}

\item{libLoc}{a string giving the location of the CoreNLP java
files. This should point to a directory which
contains, for example the file "stanford-corenlp-*.jar",
where "*" is the version number. If missing, the function
will try to find the library in the environment variable
CORENLP_HOME, and otherwise will fail.}

\item{mem}{a string giving the amount of memory to be assigned to the rJava
engine. For example, "6g" assigned 6 gigabytes of memory. At least
2 gigabytes are recommended at a minimum for running the CoreNLP
package. On a 32bit machine, where this is not possible, setting
"1800m" may also work. This option will only have an effect the first
time \code{initCoreNLP} is called, and also will not have an effect if
the java engine is already started by a seperate process.}
}
\description{
Store properties used when initalizing the CoreNLP pipeline
}

