% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{set_language}
\alias{set_language}
\title{Wrapper to set properties for a particular language}
\usage{
set_language(language)
}
\arguments{
\item{language}{a character vector describing the desired language;
should be one of: "en", "cn", "ar", "de", "fr", or "es"}
}
\description{
This function calls \code{\link{set_properties}}, setting all of the correct
properties to parse the given language. This is the most user-friendly entry
point for setting properties. The functions \code{\link{set_annotators}} and
\code{\link{set_properties}}
}
\examples{
\dontrun{
set_language("en")
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}

