% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{set_spacy_properties}
\alias{set_spacy_properties}
\title{Interface for setting up the Python pipeline}
\usage{
set_spacy_properties(entity_flag = TRUE, vector_flag = FALSE)
}
\arguments{
\item{entity_flag}{boolean. Should named entities be identified.}

\item{vector_flag}{boolean. Should word vectors be computed and saved.}
}
\description{
This function sets properties to parse text using the spaCy
python NLP pipeline. You must re-run the \code{\link{init_clean_nlp}}
for these changes to take place.
}
\examples{
\dontrun{
set_spacy_properties(vector_flag = TRUE)
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
