% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{setup_spaCy_backend}
\alias{setup_spaCy_backend}
\title{Interface for setting up the spaCy backend}
\usage{
setup_spaCy_backend(entity_flag = TRUE, vector_flag = FALSE,
  model_name = NULL)
}
\arguments{
\item{entity_flag}{boolean. Should named entities be identified.}

\item{vector_flag}{boolean. Should word vectors be computed and saved.}

\item{model_name}{string giving the model name for the Python/spaCy backend.
Defaults to English if NULL.}
}
\description{
This function sets properties to parse text using the spaCy
python NLP pipeline. You must re-run the \code{\link{init_backend}}
for these changes to take place.
}
\examples{
\dontrun{
setup_spaCy_backend(vector_flag = TRUE)
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
