% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{setup_tokenizers_backend}
\alias{setup_tokenizers_backend}
\title{Interface for setting up the tokenizers backend}
\usage{
setup_tokenizers_backend()
}
\description{
This function must be run before initializing the tokenizers
backend.
}
\examples{
\dontrun{
setup_tokenizers_backend()
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
