% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif.R
\name{tif_annotation}
\alias{tif_annotation}
\title{Wrapper for TIF Compliant Usage}
\usage{
tif_annotation(input)
}
\arguments{
\item{input}{a data frame representing a tif corpus object.
The first column should contain a unique document
identifier and the second should contain the raw
text.}
}
\value{
a data frame containing the parsed text, with one row for
         each token
}
\description{
Runs annotations over a corpus object and returns results as
a single data frame. See parameters below for the specific
input parameters. For more control, see the function
\code{\link{run_annotators}}.
}
\examples{
\dontrun{
annotation <- tif_annotation(tif_data_frame)
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
