% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{un}
\alias{un}
\title{Universal Declaration of Human Rights}
\description{
Data frame containing the 30 Articles in the United Nations'
Universal Declaration of Human Rights, ratified on 10 December 1948
in Paris, France.
}
\references{
\url{https://www.un.org/en/universal-declaration-human-rights/}
}
\keyword{data}
