% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{word_frequency}
\alias{word_frequency}
\title{Most frequent English words}
\description{
A dataset of the 150k most frequently used English words, extracted by Peter Norvig
from the Google Web Trillion Word Corpus. Frequencies are multiplied by 100.
}
\references{
\url{http://norvig.com/ngrams/}
}
\keyword{data}

