% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{write_annotation}
\alias{write_annotation}
\title{Write annotation files to disk}
\usage{
write_annotation(annotation, output_dir)
}
\arguments{
\item{annotation}{annotation file being stored}

\item{output_dir}{path to the directory where the files will be saved}
}
\description{
Takes an annotation object and stores it as a set of files in a local directory.
These are stored as plain-text csv files with column headers. To save as a compressed
format, instead directly call the function \code{\link{saveRDS}}.
}
\examples{
\dontrun{
write_annotation(annotation, "/path/to/annotation")
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}

