% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif.R
\name{cnlp_get_tif}
\alias{cnlp_get_tif}
\title{One Table Summary of an Annotation Object}
\usage{
cnlp_get_tif(annotation)
}
\arguments{
\item{annotation}{an annotation object}
}
\description{
This function pulls together several tables to provide
a one table, denormalized version of the annotation object.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
