% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cnlp_init_corenlp}
\alias{cnlp_init_corenlp}
\title{Interface for initializing the coreNLP backend}
\usage{
cnlp_init_corenlp(lang = NULL, models_dir = NULL)
}
\arguments{
\item{lang}{string giving the language name for the corenlp backend.
Defaults to "en" (English) if set to NULL.}

\item{models_dir}{directory where model files are located. Set to NULL to
use the default.}
}
\description{
This function must be run before annotating text with
the coreNLP backend. It sets the properties for the
spacy engine and loads the file using the R to Python
interface provided by reticulate.
}
\examples{
\dontrun{
cnlp_init_corenlp()
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
