% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{cnlp_init_spacy}
\alias{cnlp_init_spacy}
\title{Interface for initializing the spacy backend}
\usage{
cnlp_init_spacy(model_name = NULL)
}
\arguments{
\item{model_name}{string giving the model name for the spacy backend.
Defaults to "en" (English) if set to NULL.}
}
\description{
This function must be run before annotating text with
the spacy backend. It sets the properties for the
spacy engine and loads the file using the R to Python
interface provided by reticulate.
}
\examples{
\dontrun{
cnlp_init_spacy(vector_flag = TRUE)
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
