% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{doc_id_reset}
\alias{doc_id_reset}
\title{Reset document ids}
\usage{
doc_id_reset(annotation, start_id = 0L)
}
\arguments{
\item{annotation}{annotation object to reset the IDs of}

\item{start_id}{the starting document id. Defaults to 0.}
}
\value{
an annotation object with document ids updated across all tables.
}
\description{
Given an annotation object, this function changes all of the document ids so
that they are all contiguous integers, starting at the parameter \code{start_id}.
}
\examples{
\dontrun{
doc_id_reset(annotation, 10)
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}

