% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{NERAnnotate}
\alias{NERAnnotate}
\title{Run the annotation pipeline on a set of documents to extract entities}
\usage{
NERAnnotate(input, language = "en", lib_location = NULL, mem = "2g",
  verbose = FALSE)
}
\arguments{
\item{input}{a character vector with one document in each
element}

\item{language}{The language for CoreNLP to use or detection.}

\item{lib_location}{The path of the CoreNLP Java library on the users system.}

\item{mem}{The amount of memory that the Java CoreNLP call is allocated.
Needs to be of the form "2g" for 2GB or "500m" for 500MB.
Decimal accuracy not allowed.}

\item{verbose}{A logical value to set if Java output messages are displayed.}
}
\value{
a data.frame with the details of the detected entities.
}
\description{
Runs the entity detection algorithms from CoreNLP using either the rJava.
It initializes the CoreNLP Java object with the NER annotation parameters
and a path to a temp file that is used for processing using \code{initCoreNLPNER}.
The function returns a \code{data.frame} showing the location in the document 
where the entity occurs andthe entity type. If no entities are detected for a document 
then a row of NA values is returned.
}
