% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{read_annotation}
\alias{read_annotation}
\title{Read annotation files from disk}
\usage{
read_annotation(input_dir)
}
\arguments{
\item{input_dir}{path to the directory where the files are stored}
}
\value{
an object of class \code{annotation}
}
\description{
Loads an annotation that has been stored as a set of csv files in a local directory.
This is typically created by a call to \code{\link{annotate}} or \code{\link{write_annotation}}.
}
\examples{
\dontrun{
annotation <- read_annotation("path/to/annotation")
}

}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
